pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "new-project"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Juan Monsalve", email = "Jmonsalve@hematologico.com" },
]
classifiers = ["Programming Language :: Python :: 3.12"]
dependencies = [
  "aiohttp==3.9.5",
  "fastapi==0.111.0",
  "gunicorn==22.0.0",
  "pydantic_settings==2.2.1",
  "uvicorn==0.29.0",
  "jinja2==3.1.4",
  "weasyprint==62.3",
]

[tool.hatch.version]
path = "src/new_project/__about__.py"

[tool.hatch.envs.default.scripts]
server = "gunicorn"
dev-server = "uvicorn new_project.__init__:app --reload"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/new_project tests}"

[tool.coverage.run]
source_pkgs = ["new_project", "tests"]
branch = true
parallel = true
omit = [
  "src/new_project/__about__.py",
]

[tool.coverage.paths]
new_project = ["src/new_project", "*/new-project/src/new_project"]
tests = ["tests", "*/new-project/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
